\name{AdvancedRLab-package}
\alias{AdvancedRLab-package}
\alias{AdvancedRLab}
\docType{package}
\title{
\packageTitle{AdvancedRLab}
}
\description{
\packageDescription{AdvancedRLab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdvancedRLab}
\packageIndices{AdvancedRLab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AdvancedRLab}

Maintainer: \packageMaintainer{AdvancedRLab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
