% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a dataframe consists of three variables}

\item{init_node}{a numeric scalar for the first node which also exist in the graph}
}
\value{
The value is a vector. It is the least value (or shortest path) from the starting node to every nodes in a graph.
}
\description{
A method for finding the shortest path between the nodes in a graph.
}
\examples{
dijkstra(wiki_graph, 1) #0 7 9 20 20 11
dijkstra(wiki_graph, 3) #9 10 0 11 11 2
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra}
}

