% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{Euclidian Algorithm}
\usage{
euclidian(a, b)
}
\arguments{
\item{a}{a numeric scalar}

\item{b}{a numeric scalar}
}
\value{
The largest number that divides the two numbers (\code{a} and \code{b}) without leaving any remainder.
}
\description{
A method for finding the greatest common divisor (GCD)
}
\examples{
euclidian(123612, 13892347912) #4
euclidian(100,10000) #100
}
\references{
\url{http://en.wikipedia.org/wiki/Euclidean_algorithm}
}

