% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiki_graph-data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph}
\format{A data frame with 18 rows and 3 variables:
\describe{
 \item{v1}{from: the edges of the graph}
 \item{v2}{to: the edges of the graph}
 \item{w}{weight of the edge}
}}
\usage{
data(wiki_graph)
}
\description{
A dataset containing the nodes and the length of paths.
}
\examples{
data(wiki_graph)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra}
}
\keyword{datasets}

