\name{AdvancedRLab-package}
\alias{AdvancedRLab-package}
\alias{AdvancedRLab}
\docType{package}
\title{
\packageTitle{AdvancedRLab}
}
\description{
\packageDescription{AdvancedRLab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AdvancedRLab}
\packageIndices{AdvancedRLab}
The package contains two functions dijkstra dn euclidian. It also contains a dataset wiki_graph which can be used for the dijkstra function.
}
\author{
\packageAuthor{AdvancedRLab}

Maintainer: \packageMaintainer{AdvancedRLab}
}
\references{
http://www.ida.liu.se/~732A50/info/courseinfo.en.shtml
}
\keyword{ package }
\seealso{

}
\examples{
euclidian(10,100)
}
